Move RequestMetadata to router
