+++ python/ray/serve/handle.py
@@ -3,7 +3,7 @@ from typing import Optional
 import ray
 from ray import serve
 from ray.serve.context import TaskContext
-from ray.serve.request_params import RequestMetadata
+from ray.serve.router import RequestMetadata
 
 
 class RayServeHandle:
