+++ python/ray/serve/tests/test_backend_worker.py
@@ -8,8 +8,7 @@ from ray import serve
 import ray.serve.context as context
 from ray.serve.backend_worker import create_backend_worker, wrap_to_ray_error
 from ray.serve.controller import TrafficPolicy
-from ray.serve.request_params import RequestMetadata
-from ray.serve.router import Router
+from ray.serve.router import Router, RequestMetadata
 from ray.serve.config import BackendConfig
 from ray.serve.exceptions import RayServeException
